requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==3.0.1
email-validator==2.0.0
pillow==10.1.0
python-dotenv==1.0.0
